Move some email config to site section
On second thoughts this seems the right place for it
