+++ server/camcops_server/cc_modules/cc_constants.py
@@ -328,14 +328,17 @@ CONFIG_FILE_SERVER_SECTION = "server"
 CONFIG_FILE_EXPORT_SECTION = "export"
 
 
-class ConfigParamExportGeneral(object):
+class ConfigParamSite(object):
     """
-    Parameters allowed in the ``[export]`` section of the CamCOPS config file.
+    Parameters allowed in the main section of the CamCOPS config file.
     """
-    CELERY_BEAT_EXTRA_ARGS = "CELERY_BEAT_EXTRA_ARGS"
-    CELERY_BEAT_SCHEDULE_DATABASE = "CELERY_BEAT_SCHEDULE_DATABASE"
-    CELERY_BROKER_URL = "CELERY_BROKER_URL"
-    CELERY_WORKER_EXTRA_ARGS = "CELERY_WORKER_EXTRA_ARGS"
+    ALLOW_INSECURE_COOKIES = "ALLOW_INSECURE_COOKIES"
+    CAMCOPS_LOGO_FILE_ABSOLUTE = "CAMCOPS_LOGO_FILE_ABSOLUTE"
+    CLIENT_API_LOGLEVEL = "CLIENT_API_LOGLEVEL"
+    CTV_FILENAME_SPEC = "CTV_FILENAME_SPEC"
+    DB_URL = "DB_URL"
+    DB_ECHO = "DB_ECHO"
+    DISABLE_PASSWORD_AUTOCOMPLETE = "DISABLE_PASSWORD_AUTOCOMPLETE"
     EMAIL_FROM = "EMAIL_FROM"
     EMAIL_HOST = "EMAIL_HOST"
     EMAIL_HOST_PASSWORD = "EMAIL_HOST_PASSWORD"
@@ -344,6 +347,69 @@ class ConfigParamExportGeneral(object):
     EMAIL_REPLY_TO = "EMAIL_REPLY_TO"
     EMAIL_SENDER = "EMAIL_SENDER"
     EMAIL_USE_TLS = "EMAIL_USE_TLS"
+    EXTRA_STRING_FILES = "EXTRA_STRING_FILES"
+    LANGUAGE = "LANGUAGE"
+    LOCAL_INSTITUTION_URL = "LOCAL_INSTITUTION_URL"
+    LOCAL_LOGO_FILE_ABSOLUTE = "LOCAL_LOGO_FILE_ABSOLUTE"
+    LOCKOUT_DURATION_INCREMENT_MINUTES = "LOCKOUT_DURATION_INCREMENT_MINUTES"
+    LOCKOUT_THRESHOLD = "LOCKOUT_THRESHOLD"
+    PASSWORD_CHANGE_FREQUENCY_DAYS = "PASSWORD_CHANGE_FREQUENCY_DAYS"
+    PATIENT_SPEC = "PATIENT_SPEC"
+    PATIENT_SPEC_IF_ANONYMOUS = "PATIENT_SPEC_IF_ANONYMOUS"
+    RESTRICTED_TASKS = "RESTRICTED_TASKS"
+    SESSION_COOKIE_SECRET = "SESSION_COOKIE_SECRET"
+    SESSION_TIMEOUT_MINUTES = "SESSION_TIMEOUT_MINUTES"
+    SNOMED_TASK_XML_FILENAME = "SNOMED_TASK_XML_FILENAME"
+    SNOMED_ICD9_XML_FILENAME = "SNOMED_ICD9_XML_FILENAME"
+    SNOMED_ICD10_XML_FILENAME = "SNOMED_ICD10_XML_FILENAME"
+    TASK_FILENAME_SPEC = "TASK_FILENAME_SPEC"
+    TRACKER_FILENAME_SPEC = "TRACKER_FILENAME_SPEC"
+    WEBVIEW_LOGLEVEL = "WEBVIEW_LOGLEVEL"
+    WKHTMLTOPDF_FILENAME = "WKHTMLTOPDF_FILENAME"
+
+
+class ConfigParamServer(object):
+    """
+    Parameters allowed in the web server section of the CamCOPS config file.
+    """
+    CHERRYPY_LOG_SCREEN = "CHERRYPY_LOG_SCREEN"
+    CHERRYPY_ROOT_PATH = "CHERRYPY_ROOT_PATH"
+    CHERRYPY_SERVER_NAME = "CHERRYPY_SERVER_NAME"
+    CHERRYPY_THREADS_MAX = "CHERRYPY_THREADS_MAX"
+    CHERRYPY_THREADS_START = "CHERRYPY_THREADS_START"
+    DEBUG_REVERSE_PROXY = "DEBUG_REVERSE_PROXY"
+    DEBUG_SHOW_GUNICORN_OPTIONS = "DEBUG_SHOW_GUNICORN_OPTIONS"
+    DEBUG_TOOLBAR = "DEBUG_TOOLBAR"
+    GUNICORN_DEBUG_RELOAD = "GUNICORN_DEBUG_RELOAD"
+    GUNICORN_NUM_WORKERS = "GUNICORN_NUM_WORKERS"
+    GUNICORN_TIMEOUT_S = "GUNICORN_TIMEOUT_S"
+    HOST = "HOST"
+    PORT = "PORT"
+    PROXY_HTTP_HOST = "PROXY_HTTP_HOST"
+    PROXY_REMOTE_ADDR = "PROXY_REMOTE_ADDR"
+    PROXY_REWRITE_PATH_INFO = "PROXY_REWRITE_PATH_INFO"
+    PROXY_SCRIPT_NAME = "PROXY_SCRIPT_NAME"
+    PROXY_SERVER_NAME = "PROXY_SERVER_NAME"
+    PROXY_SERVER_PORT = "PROXY_SERVER_PORT"
+    PROXY_URL_SCHEME = "PROXY_URL_SCHEME"
+    SHOW_REQUEST_IMMEDIATELY = "SHOW_REQUEST_IMMEDIATELY"
+    SHOW_REQUESTS = "SHOW_REQUESTS"
+    SHOW_RESPONSE = "SHOW_RESPONSE"
+    SHOW_TIMING = "SHOW_TIMING"
+    SSL_CERTIFICATE = "SSL_CERTIFICATE"
+    SSL_PRIVATE_KEY = "SSL_PRIVATE_KEY"
+    TRUSTED_PROXY_HEADERS = "TRUSTED_PROXY_HEADERS"
+    UNIX_DOMAIN_SOCKET = "UNIX_DOMAIN_SOCKET"
+
+
+class ConfigParamExportGeneral(object):
+    """
+    Parameters allowed in the ``[export]`` section of the CamCOPS config file.
+    """
+    CELERY_BEAT_EXTRA_ARGS = "CELERY_BEAT_EXTRA_ARGS"
+    CELERY_BEAT_SCHEDULE_DATABASE = "CELERY_BEAT_SCHEDULE_DATABASE"
+    CELERY_BROKER_URL = "CELERY_BROKER_URL"
+    CELERY_WORKER_EXTRA_ARGS = "CELERY_WORKER_EXTRA_ARGS"
     EXPORT_LOCKDIR = "EXPORT_LOCKDIR"
     RECIPIENTS = "RECIPIENTS"
     SCHEDULE = "SCHEDULE"
