+++ server/camcops_server/cc_modules/cc_exportmodels.py
@@ -65,12 +65,14 @@ from sqlalchemy.sql.sqltypes import (
     UnicodeText,
 )
 
+from camcops_server.cc_modules.cc_constants import (
+    ConfigParamExportRecipient,
+)
 from camcops_server.cc_modules.cc_email import Email
 from camcops_server.cc_modules.cc_exportrecipient import (
     ExportRecipient,
 )
 from camcops_server.cc_modules.cc_exportrecipientinfo import (
-    ConfigParamExportRecipient,
     ExportTransmissionMethod,
 )
 from camcops_server.cc_modules.cc_filename import (
